Fallback for individual eva-course-identifier
If EVASYS_COURSE_IDENTIFIER is set in Studip config,
it may be set to a field which is empty in certain seminars.
In this case the internal studip seminar_id should be
used as fallback. Otherwise the transfer to evasys would
fail since the unique identifier for the seminar is empty.